requirements.txt
streamlit
PyPDF2
sentence-transformers
faiss-cpu
spacy
google-generativeai
PyMuPDF